Move ToU price generation to separate class
